==> Programming/Digital House (Python)/Project 2 - Real estate/3) Comparison.ipynb <==
# 1) Preparación previa

### Carga de librerías

import pandas as pd
import numpy as np
import re
from sklearn import linear_model
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error
from sklearn.model_selection import train_test_split
%matplotlib inline
from matplotlib import pyplot as plt
import seaborn as sns
import scipy as sp
from sklearn.preprocessing import PolynomialFeatures
from sklearn.model_selection import train_test_split

### Lectura del data set con dummies

data_con_dummies = pd.read_csv("data_con_dummies.csv", sep = ";")
data_con_dummies.head(5)

# 2) Gráficos

Se trabajará con distintos subconjuntos de datos

### Datos de PHs y departamentos en Palermo

dataset_test_Palermo = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Palermo', 
                                 'prop_PH', 'prop_apartment', 'surface_covered_in_m2', 
                                 'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 
                                 'dummy_amenities', 'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Palermo.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Palermo = dataset_test_Palermo[dataset_test_Palermo.ambientes_imputados == 0].copy()
X_test_Palermo = dataset_test_Palermo[dataset_test_Palermo.ambientes_train == 0].copy()

Y_train_Palermo = X_train_Palermo["precio_usd_por_m2"]
Y_test_Palermo = X_test_Palermo["precio_usd_por_m2"]

X_train_Palermo.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Palermo.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Palermo = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Palermo = lm_Palermo.fit(X_train_Palermo, Y_train_Palermo)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Palermo = lm_Palermo.predict(X_test_Palermo)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Palermo.intercept_)
print ('RM =', model_Palermo.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Palermo = model_Palermo.score(X_train_Palermo, Y_train_Palermo)
R2_train_Palermo

#media del precio_usd_por_m2
dataset_test_Palermo.precio_usd_por_m2.mean()

# Error absoluto de la media
MAE_Test_Palermo = mean_absolute_error(Y_test_Palermo, predictions_Palermo)
MAE_Test_Palermo

# Raiz cuadrada de la media del error
MSE_Test_Palermo = mean_squared_error(Y_test_Palermo, predictions_Palermo)
MSE_Test_Palermo

RMSE_Test_Palermo = np.sqrt(mean_squared_error(Y_test_Palermo, predictions_Palermo))
RMSE_Test_Palermo

# Carga de los datos de la regresión lineal simple en base a la superficie obtenida en la notebook 2

revision_datos = [ {'subconjunto': 'Superficie Cubierta',
    'R2_train': 0.0838,
    'MAE': 630.7234,
    'MSE': 616548.9624,
    'RMSE': 785.2063}]

# Se procede a agregar los nuevos datos obtenidos. Se analizarán al final

revision_datos.append({
    'subconjunto': 'PHs y departamentos en Palermo',
    'R2_train': R2_train_Palermo.round(4),
    'MAE': MAE_Test_Palermo.round(4),
    'MSE': MSE_Test_Palermo.round(4),
    'RMSE': RMSE_Test_Palermo.round(4),
})

dataset_test_Palermo.columns

## Departamentos en Rosario

dataset_test_Rosario = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Rosario', 
                                   'prop_apartment', 'surface_covered_in_m2', 
                                   'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 'dummy_amenities', 
                                   'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Rosario.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Rosario = dataset_test_Rosario[dataset_test_Rosario.ambientes_imputados == 0].copy()
X_test_Rosario = dataset_test_Rosario[dataset_test_Rosario.ambientes_train == 0].copy()

Y_train_Rosario = X_train_Rosario["precio_usd_por_m2"]
Y_test_Rosario = X_test_Rosario["precio_usd_por_m2"]

X_train_Rosario.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Rosario.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Rosario = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Rosario = lm_Rosario.fit(X_train_Rosario, Y_train_Rosario)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Rosario = lm_Rosario.predict(X_test_Rosario)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Rosario.intercept_)
print ('RM =', model_Rosario.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Rosario = model_Rosario.score(X_train_Rosario, Y_train_Rosario)
R2_train_Rosario

#media del precio_usd_por_m2
dataset_test_Rosario.precio_usd_por_m2.mean()


# Error absoluto de la media
MAE_Test_Rosario = mean_absolute_error(Y_test_Rosario, predictions_Rosario)
MAE_Test_Rosario

# Raiz cuadrada de la media del error
MSE_Test_Rosario = mean_squared_error(Y_test_Rosario, predictions_Rosario)
MSE_Test_Rosario

RMSE_Test_Rosario = np.sqrt(mean_squared_error(Y_test_Rosario, predictions_Rosario))
RMSE_Test_Rosario

revision_datos.append({
    'subconjunto': 'Departamentos en Rosario',
    'R2_train': R2_train_Rosario.round(4),
    'MAE': MAE_Test_Rosario.round(4),
    'MSE': MSE_Test_Rosario.round(4),
    'RMSE': RMSE_Test_Rosario.round(4),
})

data_con_dummies.columns

## Departamentos en Córdoba

dataset_test_Cordoba = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Córdoba', 
                                   'prop_apartment', 'surface_covered_in_m2', 
                                   'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 'dummy_amenities', 
                                   'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Cordoba.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Cordoba = dataset_test_Cordoba[dataset_test_Cordoba.ambientes_imputados == 0].copy()
X_test_Cordoba = dataset_test_Cordoba[dataset_test_Cordoba.ambientes_train == 0].copy()

Y_train_Cordoba = X_train_Cordoba["precio_usd_por_m2"]
Y_test_Cordoba = X_test_Cordoba["precio_usd_por_m2"]

X_train_Cordoba.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Cordoba.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Cordoba = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Cordoba = lm_Cordoba.fit(X_train_Cordoba, Y_train_Cordoba)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Cordoba = lm_Cordoba.predict(X_test_Cordoba)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Cordoba.intercept_)
print ('RM =', model_Cordoba.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Cordoba = model_Cordoba.score(X_train_Cordoba, Y_train_Cordoba)
R2_train_Cordoba

#media del precio_usd_por_m2
dataset_test_Cordoba.precio_usd_por_m2.mean()


# Error absoluto de la media
MAE_Test_Cordoba = mean_absolute_error(Y_test_Cordoba, predictions_Cordoba)
MAE_Test_Cordoba

# Raiz cuadrada de la media del error
MSE_Test_Cordoba = mean_squared_error(Y_test_Cordoba, predictions_Cordoba)
MSE_Test_Cordoba

RMSE_Test_Cordoba = np.sqrt(mean_squared_error(Y_test_Cordoba, predictions_Cordoba))
RMSE_Test_Cordoba

revision_datos.append({
    'subconjunto': 'Departamentos en Córdoba',
    'R2_train': R2_train_Cordoba.round(4),
    'MAE': MAE_Test_Cordoba.round(4),
    'MSE': MSE_Test_Cordoba.round(4),
    'RMSE': RMSE_Test_Cordoba.round(4),
})

## PHs y casas en Pinamar

dataset_test_Pinamar = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Pinamar', 
                                   'prop_PH', 'prop_house','surface_covered_in_m2', 
                                   'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 'dummy_amenities', 
                                   'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Pinamar.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Pinamar = dataset_test_Pinamar[dataset_test_Pinamar.ambientes_imputados == 0].copy()
X_test_Pinamar = dataset_test_Pinamar[dataset_test_Pinamar.ambientes_train == 0].copy()

Y_train_Pinamar = X_train_Pinamar["precio_usd_por_m2"]
Y_test_Pinamar = X_test_Pinamar["precio_usd_por_m2"]

X_train_Pinamar.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Pinamar.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Pinamar = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Pinamar = lm_Pinamar.fit(X_train_Pinamar, Y_train_Pinamar)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Pinamar = lm_Pinamar.predict(X_test_Pinamar)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Pinamar.intercept_)
print ('RM =', model_Pinamar.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Pinamar = model_Pinamar.score(X_train_Pinamar, Y_train_Pinamar)
R2_train_Pinamar

#media del precio_usd_por_m2
dataset_test_Pinamar.precio_usd_por_m2.mean()

# Error absoluto de la media
MAE_Test_Pinamar = mean_absolute_error(Y_test_Pinamar, predictions_Pinamar)
MAE_Test_Pinamar

# Raiz cuadrada de la media del error
MSE_Test_Pinamar = mean_squared_error(Y_test_Pinamar, predictions_Pinamar)
MSE_Test_Pinamar

RMSE_Test_Pinamar = np.sqrt(mean_squared_error(Y_test_Pinamar, predictions_Pinamar))
RMSE_Test_Pinamar

revision_datos.append({
    'subconjunto': 'PHs y casas en Pinamar',
    'R2_train': R2_train_Pinamar.round(4),
    'MAE': MAE_Test_Pinamar.round(4),
    'MSE': MSE_Test_Pinamar.round(4),
    'RMSE': RMSE_Test_Pinamar.round(4),
})

## Propiedades en Mar del Plata

dataset_test_Mardel = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Mar del Plata', 
                                   'prop_PH', 'prop_apartment', 'prop_house','surface_covered_in_m2', 
                                   'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 'dummy_amenities', 
                                   'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Mardel.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Mardel = dataset_test_Mardel[dataset_test_Mardel.ambientes_imputados == 0].copy()
X_test_Mardel = dataset_test_Mardel[dataset_test_Mardel.ambientes_train == 0].copy()

Y_train_Mardel = X_train_Mardel["precio_usd_por_m2"]
Y_test_Mardel = X_test_Mardel["precio_usd_por_m2"]

X_train_Mardel.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Mardel.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Mardel = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Mardel = lm_Mardel.fit(X_train_Mardel, Y_train_Mardel)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Mardel = lm_Mardel.predict(X_test_Mardel)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Mardel.intercept_)
print ('RM =', model_Mardel.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Mardel = model_Mardel.score(X_train_Mardel, Y_train_Mardel)
R2_train_Mardel

#media del precio_usd_por_m2
dataset_test_Mardel.precio_usd_por_m2.mean()

# Error absoluto de la media
MAE_Test_Mardel = mean_absolute_error(Y_test_Mardel, predictions_Mardel)
MAE_Test_Mardel

# Raiz cuadrada de la media del error
MSE_Test_Mardel = mean_squared_error(Y_test_Mardel, predictions_Mardel)
MSE_Test_Mardel

RMSE_Test_Mardel = np.sqrt(mean_squared_error(Y_test_Mardel, predictions_Mardel))
RMSE_Test_Mardel

revision_datos.append({
    'subconjunto': 'Propiedades en Mar del Plata',
    'R2_train': R2_train_Mardel.round(4),
    'MAE': MAE_Test_Mardel.round(4),
    'MSE': MSE_Test_Mardel.round(4),
    'RMSE': RMSE_Test_Mardel.round(4),
})

## Casas en Punilla

dataset_test_Punilla = data_con_dummies[['ambientes_imputados','ambientes_train','partido_Punilla', 
                                   'prop_house','surface_covered_in_m2', 
                                   'amb_1.0', 'amb_2.0', 'amb_3.0', 'amb_4.0', 'dummy_amenities', 
                                   'precio_usd_por_m2']].copy()

plt.figure(figsize=(10,7))
sns.heatmap(dataset_test_Punilla.corr(),annot=True, cmap='Reds')
plt.title("Correlation Heatmap")
plt.show()

X_train_Punilla = dataset_test_Punilla[dataset_test_Punilla.ambientes_imputados == 0].copy()
X_test_Punilla = dataset_test_Punilla[dataset_test_Punilla.ambientes_train == 0].copy()

Y_train_Punilla = X_train_Punilla["precio_usd_por_m2"]
Y_test_Punilla = X_test_Punilla["precio_usd_por_m2"]

X_train_Punilla.drop(["precio_usd_por_m2"], axis = 1, inplace = True)
X_test_Punilla.drop(["precio_usd_por_m2"], axis = 1, inplace = True)

lm_Punilla = linear_model.LinearRegression()
# Fiteamos el modelo sobre los vectores X e Y.
model_Punilla = lm_Punilla.fit(X_train_Punilla, Y_train_Punilla)

# Guardamos  las predicciones en un nuevo vector que llamaremos predictions.
predictions_Punilla = lm_Punilla.predict(X_test_Punilla)

# Imprimimos el intercepto y los coeficientes como atributos del objeto entrenado.
print ('Intercepto =', model_Punilla.intercept_)
print ('RM =', model_Punilla.coef_)

# Imprimos la metrica que mide la bondad de ajusto del modelo. En este caso el R2.
R2_train_Punilla = model_Punilla.score(X_train_Punilla, Y_train_Punilla)
R2_train_Punilla

#media del precio_usd_por_m2
dataset_test_Punilla.precio_usd_por_m2.mean()


# Error absoluto de la media
MAE_Test_Punilla = mean_absolute_error(Y_test_Punilla, predictions_Punilla)
MAE_Test_Punilla

# Raiz cuadrada de la media del error
MSE_Test_Punilla = mean_squared_error(Y_test_Punilla, predictions_Punilla)
MSE_Test_Punilla

RMSE_Test_Punilla = np.sqrt(mean_squared_error(Y_test_Punilla, predictions_Punilla))
RMSE_Test_Punilla

revision_datos.append({
    'subconjunto': 'Casas en Punilla',
    'R2_train': R2_train_Punilla.round(4),
    'MAE': MAE_Test_Punilla.round(4),
    'MSE': MSE_Test_Punilla.round(4),
    'RMSE': RMSE_Test_Punilla.round(4),
})

# 3) Resumen de datos

revision_datos
